LayoutTests/platform/ios/imported/w3c/web-platform-tests/screen-orientation/lock-sandboxed-iframe-expected.txt

CONSOLE MESSAGE: Error while parsing the 'sandbox' attribute: 'allow-orientation-lock' is an invalid sandbox flag.

FAIL lock-sandboxed-iframe assert_true: iframe supports 'allow-orientation-lock' expected true got false
FAIL Test without 'allow-orientation-lock' sandboxing directive assert_equals: screen.lockOrientation() throws a SecurityError expected "SecurityError" but got "portrait-primary"
PASS Test with 'allow-orientation-lock' sandboxing directive
